(refactor): Improve folder structure
