+++ src/odds/adapters/odds.adapter.ts
@@ -1,6 +1,6 @@
 import { injectable } from "inversify";
-import { IOddsService } from "./odds.port";
-import { Odd } from "./odds.model";
+import { IOddsService } from "../ports/odds.port";
+import { Odd } from "../models/odds.model";
 
 @injectable()
 export class OddsAdapter implements IOddsService {
