+++ src/users/models/users.model.ts
